requirements.txt
Mastodon.py==1.8.1
requests==2.25.1
requests-toolbelt==0.9.1
